+++ src/pages/api/(templates)/template.ts
@@ -1,5 +1,5 @@
 import { imageTemplate as Mediatemp } from "@/components/Carousel/stories"
-import { template as DStemp } from "@/components/DonateSection/stories"
 import { template as NLtemp } from "@/components/Newsletter/stories"
+import { template as DStemp } from "@/components/SectionDonate/stories"
 
 export { Mediatemp, DStemp, NLtemp }
